refactor: move test package from module
