+++ tests/test_aioimaplib_capabilities.py
@@ -5,8 +5,8 @@ import pytest
 
 from aioimaplib import aioimaplib
 from aioimaplib.aioimaplib import Abort
-from aioimaplib.tests.imapserver import ImapProtocol, ServerState, MockImapServer
-from aioimaplib.tests.server_fixture import with_server, login_user_async, main_test
+from aioimaplib.imapserver import ImapProtocol, ServerState, MockImapServer
+from tests.server_fixture import with_server, login_user_async, main_test
 
 aioimaplib.log.setLevel(logging.WARNING)
 sh = logging.StreamHandler()
